requirements.txt
Flask==2.3.3
requests==2.32.0
spotipy==2.25.1
python-dotenv==1.0.0
plotly==5.18.0
